Completed exercises 1-4
I was able to completed all of the exercise; however, I was only able to
complete the first test or exercise 4. I looked at the solutions and
honestly, it is beyond my current understanding.
I'll come back to it and completed the extra tests but for now I'm going
to continue with exercisde 5.