Completed exercises 8-10
The calculator and palindromes exercises were pretty straightforward.
However, the fibonacci exercise was a bit more challenging, because I
didn't understand the algortihm at first. I had to look it up and to
be honest, I still don't get it. However, I was able to easily
understand how to find and locate the element in the array. With the
last two test, for some reason when entering a number that is in a
string, I wild and crazy numbers. Not sure why that is happening, but
at this moment, I'm going to move on to the next exercise.