Completed 11_getTheTitles
This one was a bit difficult for me because I didn't fully understand
the map function. I kinda understand but I barely know how to use it. I
need to get more practice with this.

Oh and btw I'm not going to do exercise 12. I completely don't
understand reduce(), new Date(), or exactly how to calculate the birth
and death and put it into a lopp and identify which one is the oldest.